Checkpoint of my BirdsEyeView tracking doc.
This may be of interest to me only, or it might be a generally
helpful way to keep our collective eye on a bigger-picture.
NB: the biggest picture is probably Robot Strategy.  This is
not intended to represent that. Rather it represents how we're
serving the current robot strategy by implementing/prioritizing
subsystems within the programming team.

This is a work-in-progress and definitely not authoritative.