Fix a few gradle deprecation warnings.
The following currently exists:

gradle Using insecure protocols with repositories has been deprecated.

and appears to be due to the fact that the rev-robotics repos are http
and not https.  There is a way to express that this is acceptable on
a per-repo basis, but I didn't have the time to figure it out.